Adding a fix for some compilers which think that a uid_t could be ambiguously converted to a string (by way of several of the numeric types).

+++ src/common/OW_UserUtils.cpp
@@ -55,7 +55,7 @@ String getEffectiveUserId()
 #ifdef OW_WIN32
 #error "port me!"
 #else
-	return String(::geteuid());
+	return String(Int64(::geteuid()));
 #endif
 }
